Split interaction query module and fix timestamp filtering
