+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ### Fixed
 
+- **Interaction GraphQL date filters now treat equivalent UTC offsets as the same instant**: interaction browse and search filters no longer fall back to raw string ordering after successfully parsing RFC 3339 timestamps. This fixes inclusive `since` / `until` boundaries when GraphQL normalises `Z` inputs to `+00:00`, restoring slim and global `interactionSessions` queries that would otherwise drop matching sessions at the time-window edge.
 - **Codex interaction and checkpoint persistence now survive missing `Stop.transcriptPath` payloads**: the lifecycle-routed Codex hook path now resolves real Codex rollout transcripts from saved pre-prompt/session state and the date-sharded `~/.codex/sessions` store before failing turn-end capture. This restores persistence of `interaction_sessions`, `interaction_turns`, `interaction_events`, and checkpoint save-step data for Codex sessions, keeps later turns scoped to their new transcript fragment instead of replaying from offset `0`, and teaches transcript metadata extraction to read real Codex `response_item` payloads for prompt and summary capture.
 - **Lifecycle token usage now persists for Claude Code, Codex, and OpenCode sessions**: Claude Code and OpenCode lifecycle routing now expose their existing transcript-backed token calculators, so turn-end capture can persist token usage into checkpoint save-step metadata, `interaction_turns`, and the downstream Event DB sync instead of leaving relational and event-side token fields empty. Codex token accounting now computes per-turn deltas from cumulative `event_msg` `token_count` records, separates cached prompt tokens from regular input tokens, and counts multiple model calls within the same turn correctly. Gemini CLI and Copilot CLI remain covered by lifecycle token capture, while Cursor still cannot emit token usage until upstream adds token counters to its local session artefacts.
